+++ query/src/sql/optimizer/pattern_extractor.rs
@@ -128,114 +128,3 @@ impl PatternExtractor {
         results
     }
 }
-
-#[cfg(test)]
-mod test {
-    use std::rc::Rc;
-
-    use crate::sql::optimizer::m_expr::MExpr;
-    use crate::sql::optimizer::memo::Memo;
-    use crate::sql::optimizer::pattern_extractor::PatternExtractor;
-    use crate::sql::optimizer::SExpr;
-    use crate::sql::LogicalGet;
-    use crate::sql::LogicalProject;
-    use crate::sql::Plan;
-
-    #[test]
-    fn test_unary_expression() {
-        // LogicalProject
-        // \
-        //  LogicalGet
-        let expr = SExpr::create_unary(
-            Rc::new(Plan::LogicalProject(LogicalProject::default())),
-            SExpr::create_leaf(Rc::new(Plan::LogicalGet(LogicalGet::default()))),
-        );
-
-        // LogicalProject
-        // \
-        //  Pattern
-        let pattern = SExpr::create_unary(
-            Rc::new(Plan::LogicalProject(Default::default())),
-            SExpr::create_leaf(Rc::new(Plan::Pattern)),
-        );
-
-        let mut pattern_extractor = PatternExtractor::create();
-        let mut memo = Memo::create();
-        memo.init(expr.clone()).unwrap();
-
-        let group_expression = memo
-            .root()
-            .unwrap()
-            .iter()
-            .take(1)
-            .cloned()
-            .collect::<Vec<MExpr>>()[0]
-            .clone();
-        let result = pattern_extractor.extract(&memo, &group_expression, &pattern);
-
-        let expected = vec![SExpr::create(
-            Rc::new(Plan::LogicalProject(LogicalProject::default())),
-            vec![SExpr::create(
-                Rc::new(Plan::LogicalGet(LogicalGet::default())),
-                vec![],
-                Some(0),
-            )],
-            Some(1),
-        )];
-        assert_eq!(result, expected);
-    }
-
-    #[test]
-    fn test_multiple_expression() {
-        // LogicalProject
-        // \
-        //  LogicalGet
-        let expr = SExpr::create_unary(
-            Rc::new(Plan::LogicalProject(LogicalProject::default())),
-            SExpr::create_leaf(Rc::new(Plan::LogicalGet(LogicalGet::default()))),
-        );
-
-        // LogicalProject
-        // \
-        //  LogicalGet
-        let pattern = SExpr::create_unary(
-            Rc::new(Plan::LogicalProject(Default::default())),
-            SExpr::create_leaf(Rc::new(Plan::LogicalGet(LogicalGet::default()))),
-        );
-
-        let mut pattern_extractor = PatternExtractor::create();
-        let mut memo = Memo::create();
-        memo.init(expr.clone()).unwrap();
-
-        memo.insert_m_expr(
-            0,
-            MExpr::create(0, Rc::new(Plan::LogicalGet(LogicalGet::default())), vec![]),
-        )
-        .unwrap();
-
-        println!("{:?}", memo);
-
-        let group_expression = memo
-            .root()
-            .unwrap()
-            .iter()
-            .take(1)
-            .cloned()
-            .collect::<Vec<MExpr>>()[0]
-            .clone();
-        let result = pattern_extractor.extract(&memo, &group_expression, &pattern);
-
-        let expected_expr = SExpr::create(
-            Rc::new(Plan::LogicalProject(LogicalProject::default())),
-            vec![SExpr::create(
-                Rc::new(Plan::LogicalGet(LogicalGet::default())),
-                vec![],
-                Some(0),
-            )],
-            Some(1),
-        );
-
-        let expected = vec![expected_expr.clone(), expected_expr.clone()];
-        assert_eq!(result, expected);
-    }
-}
