+++ query/tests/it/sql/mod.rs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+mod optimizer;
 mod parser;
 mod plan_parser;
 mod sql_parser;
